// src/types/interfaces.d.ts
export interface IClientOptions {
    host: string;
    apiHost?: string;
    authToken?: string;
    jwt?: string;
    jid?: string;
    jidResource?: string;
    reconnectOnNoLongerSubscribed?: boolean;
    optOutOfWebrtcStatsTelemetry?: boolean;
    allowIPv6?: boolean;
    logger?: ILogger;
    logLevel?: LogLevel;
    logFormatters?: LogFormatterFn[];
    signalIceConnected?: boolean;
    useServerSidePings?: boolean;
    appName?: string;
    appVersion?: string;
    appId?: string;
    customHeaders?: ICustomHeader;
    /** Allow bulk topic resubscribe to succeed or fail per-topic rather than all or nothing */
    enablePartialBulkResubscribe?: boolean;
    /** Genesys internal use only - non-Genesys apps that pass in `alertableInteractionTypes` may experience unexpected behavior */
    alertableInteractionTypes?: AlertableInteractionTypes[];
}
export interface ICustomHeader {
    [header: string]: string;
}
export declare enum AlertableInteractionTypes {
    voice = "voice"
}
// ...
export interface ExtendedRTCIceServer extends RTCIceServer {
    type: string;
}
// ...
export type SessionTypesAsStrings = 'softphone' | 'screenShare' | 'screenRecording' | 'liveScreenMonitoring' | 'collaborateVideo' | 'unknown';
export declare enum SessionTypes {
    softphone = "softphone",
    collaborateVideo = "collaborateVideo",
    acdScreenShare = "screenShare",
    screenRecording = "screenRecording",
    liveScreenMonitoring = "liveScreenMonitoring",
    unknown = "unknown"
}
export interface ISessionInfo extends IPendingSession {
}
export interface IPendingSession {
    sessionId: string;
    id: string;
    autoAnswer: boolean;
    toJid: string;
    fromJid: string;
    conversationId: string;
    originalRoomJid?: string;
    sdpOverXmpp?: boolean;
    privAnswerMode?: 'Auto';
    fromUserId?: string;
    roomJid?: string;
    accepted?: boolean;
    meetingId?: string;
    sessionType: SessionTypes | SessionTypesAsStrings;
}